Add a "scripts" folder to hold oneshot scripts, or unused scripts.
This commit does not intend on being exhaustive but at the very least it
starts something that can be completed later.
